navbar - separate concerns
