+++ app/components/header/brand.tsx
@@ -1,0 +1,5 @@
+export default function NavbarBrand() {
+  return (
+    <a className="btn btn-ghost text-xl">ranasalem.io</a>
+  )
+}
