+++ app/components/header/links.tsx
@@ -1,0 +1,11 @@
+export default function NavLinks() {
+  return (
+    <ul className="menu menu-horizontal px-1">
+      <li><a>About</a></li>
+      <li><a>Projects</a></li>
+      <li><a>Blog</a></li>
+      <li><a>Toolbox</a></li>
+      <li><a>Contact</a></li>
+    </ul>
+  )
+}
