+++ app/page.tsx
@@ -1,7 +1,7 @@
-import Navbar from "./components/navbar";
-
 export default function Home() {
   return (
-    <Navbar />
+    <div>
+      <h1>hello world</h1>
+    </div>
   );
 }
